requirements.txt
google-auth
google-cloud-aiplatform
google_cloud_pipeline_components
google-cloud-storage
kfp
matplotlib
numpy
pandas
